QAA-2471: Slack bot doesn't work with schedule
Slack bot wasn't working properly with scheduled runs. This is due to the fact that on scheduled runs, the Github context is different from during a repo dispatch. This causes the retrieved repo link to be undefined, which Slack checks for and then returns an error 400 for.

This change is to add an additional clause to the or to build the link manually if the other options aren't available.